refactor(agent): convert keypair service to api
